requirements.txt
irc
sh
daemon
validate_email
pyyaml
mako
